Fixed bug 205.
http://dev.bukkit.org/bukkit-plugins/statistics/tickets/205-not-showing-all-online-players/

If a player gets removed from the database we also have to remove the metadata from the player otherwise the plugin tries to use an outdated (not existing) player Id if the player joins again after he got removed from the database. Due to the fact that bukkit only allows Metadata changes on online players the /stats remove <player> command will probably lead to problems in the future if someone removes an player form the database and the player is joining the server afterwards.